(#117) (#45) address review comments
